requirements.txt
Django==4.2.13
djangorestframework==3.15.2
psycopg2-binary==2.9.9
Pillow==10.4.0
django-cors-headers==4.4.0
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.1
ipython==8.18.1
django-filter==24.3
djoser==2.2.3
python-dotenv==1.0.1